+++ tests/eko/test_ev_op_flavors.py
