+++ tests/eko/test_kernels_ei.py
