+++ tests/eko/test_kernels_ns.py
