+++ tests/eko/test_kernels_s.py
