+++ tests/eko/test_strong_coupling.py
